+++ src/main/java/ru/yandex/practicum/filmorate/model/User.java
@@ -14,7 +14,7 @@ import java.util.Set;
 @Data
 @UserNameConstraint
 public class User {
-    private Integer id;
+    private long id;
     @NotBlank(message = "Email can't be null or empty")
     @Email(message = "Incorrect email")
     private final String email;
@@ -23,5 +23,5 @@ public class User {
     private String name;
     @Past(message = "Birthday must be in past")
     private final LocalDate birthday;
-    private Set<Integer> friends = new HashSet<>();
+    private Set<Long> friends = new HashSet<>();
 }
